+++ Thrive.csproj
@@ -202,7 +202,6 @@
     <Compile Include="src\saving\SaveLoadable.cs" />
     <Compile Include="src\saving\SaveContext.cs" />
     <Compile Include="src\saving\SaveHelper.cs" />
-    <Compile Include="src\saving\SaveManager.cs" />
     <Compile Include="src\saving\SaveManagerGUI.cs" />
     <Compile Include="src\saving\SaveStatusOverlay.cs" />
     <Compile Include="src\saving\serializers\Base64BinaryConverter.cs" />
