+++ src/saving/SaveHelper.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Diagnostics;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Godot;
@@ -9,6 +10,12 @@ using Godot;
 /// </summary>
 public static class SaveHelper
 {
+    public enum SaveOrder
+    {
+        LastModifiedFirst,
+        FileSystem,
+    }
+
     /// <summary>
     ///   Quick save from the microbe stage
     /// </summary>
@@ -91,7 +98,7 @@ public static class SaveHelper
     public static void QuickLoad()
     {
         // TODO: is there a way to to find the latest modified file without checking them all?
-        var save = SaveManager.CreateListOfSaves(SaveManager.SaveOrder.LastModifiedFirst).FirstOrDefault();
+        var save = CreateListOfSaves(SaveOrder.LastModifiedFirst).FirstOrDefault();
 
         if (save == null)
         {
@@ -168,6 +175,84 @@ public static class SaveHelper
     /// </summary>
     private static void QueueRemoveExcessQuickSaves()
     {
-        TaskExecutor.Instance.AddTask(new Task(SaveManager.RemoveExcessQuickSaves));
+        TaskExecutor.Instance.AddTask(new Task(RemoveExcessQuickSaves));
+    }
+
+    public static void RemoveExcessQuickSaves()
+    {
+        // throw new System.NotImplementedException();
+    }
+
+    /// <summary>
+    ///   Returns a list of all saves
+    /// </summary>
+    /// <returns>The list of save names</returns>
+    public static List<string> CreateListOfSaves(SaveOrder order = SaveOrder.LastModifiedFirst)
+    {
+        var result = new List<string>();
+
+        using (var directory = new Directory())
+        {
+            if (!directory.DirExists(Constants.SAVE_FOLDER))
+                return result;
+
+            directory.Open(Constants.SAVE_FOLDER);
+            directory.ListDirBegin(true, true);
+
+            while (true)
+            {
+                var filename = directory.GetNext();
+
+                if (string.IsNullOrEmpty(filename))
+                    break;
+
+                if (!filename.EndsWith(Constants.SAVE_EXTENSION, StringComparison.Ordinal))
+                    continue;
+
+                result.Add(filename);
+            }
+
+            directory.ListDirEnd();
+        }
+
+        switch (order)
+        {
+            case SaveOrder.LastModifiedFirst:
+                {
+                    using (var file = new File())
+                    {
+                        result = result.OrderByDescending(item =>
+                            file.GetModifiedTime(PathUtils.Join(Constants.SAVE_FOLDER, item))).ToList();
+                    }
+
+                    break;
+                }
+
+            default:
+                break;
+        }
+
+        return result;
+    }
+
+    /// <summary>
+    ///   Counts the total number of saves and how many bytes they take up
+    /// </summary>
+    public static (int count, long diskSpace) CountSaves()
+    {
+        int count = 0;
+        long totalSize = 0;
+
+        using (var file = new File())
+        {
+            foreach (var save in CreateListOfSaves())
+            {
+                file.Open(PathUtils.Join(Constants.SAVE_FOLDER, save), File.ModeFlags.Read);
+                ++count;
+                totalSize += file.GetLen();
+            }
+        }
+
+        return (count, totalSize);
     }
 }
