+++ src/saving/SaveList.cs
@@ -113,7 +113,7 @@ public class SaveList : ScrollContainer
         }
 
         loadingItem.Visible = true;
-        readSavesList = new Task<List<string>>(() => SaveManager.CreateListOfSaves());
+        readSavesList = new Task<List<string>>(() => SaveHelper.CreateListOfSaves());
         TaskExecutor.Instance.AddTask(readSavesList);
     }
